backend/requirements.txt
# API and Authentication
smartapi-python==1.5.5
pyotp==2.9.0
python-dotenv==1.0.1
websocket-client==1.7.0

# Database
duckdb==0.9.2

# Utilities
logzero==1.7.0
pytz==2024.1
pandas==2.2.0  # Required by smart-api-python for historical data

# Technical Analysis
pandas-ta==0.3.14b

# Type Hints
typing-extensions==4.9.0

# Added from the code block
pyarrow==14.0.1
numpy==1.26.4

# FastAPI and Dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
httpx==0.27.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5 